Fix all pd.NA boolean comparison errors across app
Add safe_val() helper that converts pd.NA/NaN/None to float defaults.
Apply it to all dial_scores construction (budget optimizer, playbook,
PDF report). Fix classify_score in playbook.py. Remove Int64 nullable
dtype from liquidity_pct that was producing pd.NA instead of NaN.

+++ irci/liquidity.py
@@ -230,7 +230,7 @@ def add_liquidity_percentile(
             out["quarter_end"] = x["quarter_end"]
         else:
             out["quarter_end"] = x.index.get_level_values("quarter_end") if "quarter_end" in x.index.names else x.index
-        out["liquidity_pct"] = dial.round().astype("Int64")
+        out["liquidity_pct"] = dial.round()
         return out
 
     res = (
